+++ ms-registrations/src/main/java/com/juancastellano/msregistrations/repository/RegistrationRepository.java
@@ -8,16 +8,16 @@ import org.springframework.stereotype.Repository;
 import com.juancastellano.msregistrations.entity.Registration;
 
 @Repository
-public interface RegistrationRepository extends JpaRepository<Registration, Integer> {
+public interface RegistrationRepository extends JpaRepository<Registration, Long> {
 
-    Registration findByStudentIdAndCourseId(int studentId, int courseId);
+    Registration findByStudentIdAndCourseId(Long studentId, Long courseId);
 
-    void deleteAllByCourseId(int courseId);
+    void deleteAllByCourseId(Long courseId);
 
-    void deleteAllByStudentId(int studentId);
+    void deleteAllByStudentId(Long studentId);
 
-    List<Registration> findAllByStudentId(int studentId);
+    List<Registration> findAllByStudentId(Long studentId);
 
-    List<Registration> findAllByCourseId(int courseId);
+    List<Registration> findAllByCourseId(Long courseId);
 
 }
